Refactor view models and update bindings
- Replaced `BowlerListOverviewViewModel` with `BowlerOverviewViewModel` in `MauiProgram`.
- Updated namespaces for view models from `HangTab.ViewModels` to `HangTab.ViewModels.Items`.
- Renamed `BowlerListOverviewViewModel` to `BowlerOverviewViewModel` and updated related XAML files.
- Adjusted XAML files for `BowlerOverviewPage`, `BowlerSelectSubPage`, `CurrentWeekOverviewPage`, and `SeasonOverviewPage` to reference new view model types.
- Updated `DataTemplate` bindings in XAML to reflect new item view model types.
- Aligned `BowlerOverviewPage` constructor with the new view model structure.
- Overall, these changes improve the organization and maintainability of the codebase.
