+++ HangTab/MauiProgram.cs
@@ -84,7 +84,7 @@ public static class MauiProgram
 
     private static MauiAppBuilder RegisterViewModels(this MauiAppBuilder builder)
     {
-        builder.Services.AddSingleton<BowlerListOverviewViewModel>();
+        builder.Services.AddSingleton<BowlerOverviewViewModel>();
         builder.Services.AddSingleton<CurrentWeekOverviewViewModel>();
         builder.Services.AddSingleton<SeasonOverviewViewModel>();
         builder.Services.AddSingleton<SettingsViewModel>();
