+++ HangTab/Services/Mappers/WeekMapper.cs
@@ -1,6 +1,7 @@
 ﻿using HangTab.Extensions;
 using HangTab.Models;
 using HangTab.ViewModels;
+using HangTab.ViewModels.Items;
 
 namespace HangTab.Services.Mappers;
 internal static class WeekMapper
