+++ HangTab/Views/BowlerOverviewPage.xaml
@@ -5,10 +5,11 @@
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     xmlns:views="clr-namespace:HangTab.Views.Base"
     xmlns:vm="clr-namespace:HangTab.ViewModels"
+    xmlns:vmitems="clr-namespace:HangTab.ViewModels.Items"
     xmlns:models="clr-namespace:HangTab.Models"
     xmlns:controls="clr-namespace:HangTab.Views.Controls"
     x:Class="HangTab.Views.BowlerOverviewPage"
-    x:DataType="vm:BowlerListOverviewViewModel">
+    x:DataType="vm:BowlerOverviewViewModel">
     
     <!--Page titlebar-->
     <Shell.TitleView>
@@ -62,7 +63,7 @@
                 SelectionChangedCommand="{Binding NavigateToEditSelectedBowlerCommand}">
                 <CollectionView.ItemTemplate>
                     <DataTemplate
-                        x:DataType="vm:BowlerListItemViewModel">
+                        x:DataType="vmitems:BowlerListItemViewModel">
                         
                         <Border
                             Margin="5,2"
