+++ HangTab/Views/BowlerSelectSubPage.xaml
@@ -5,6 +5,7 @@
     xmlns:models="clr-namespace:HangTab.Models"
     xmlns:views="clr-namespace:HangTab.Views.Base"
     xmlns:vm="clr-namespace:HangTab.ViewModels"
+    xmlns:vmitems="clr-namespace:HangTab.ViewModels.Items"
     xmlns:controls="clr-namespace:HangTab.Views.Controls"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     xmlns:behaviors="clr-namespace:HangTab.Views.Controls.Behaviors"
@@ -58,7 +59,7 @@
             
             <CollectionView.ItemTemplate>
                 <DataTemplate
-                    x:DataType="vm:SubListItemViewModel">
+                    x:DataType="vmitems:SubListItemViewModel">
                     
                     <Border
                         Margin="5,2"
