+++ HangTab/Views/CurrentWeekOverviewPage.xaml
@@ -6,6 +6,7 @@
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
     xmlns:views="clr-namespace:HangTab.Views.Base"
     xmlns:vm="clr-namespace:HangTab.ViewModels"
+    xmlns:vmitems="clr-namespace:HangTab.ViewModels.Items"
     x:Class="HangTab.Views.CurrentWeekOverviewPage"
     x:DataType="vm:CurrentWeekOverviewViewModel">
     
@@ -43,7 +44,7 @@
 
                     <CollectionView.ItemTemplate>
                         <DataTemplate
-                            x:DataType="vm:CurrentWeekListItemViewModel">
+                            x:DataType="vmitems:CurrentWeekListItemViewModel">
                             
                             <VerticalStackLayout>
                                 <SwipeView>
